requirements.txt
Django
djangorestframework
sphinx
factory-boy==2.9.2
